+++ components/common_communication/include/common_parser.h
@@ -12,7 +12,7 @@
 
 esp_err_t common_parser_string_to_config(const char* sting_in, pivot_config* config_out);
 
-esp_err_t common_parser_string_to_gnss(const char* sting_in, uint8_t* angle, time_t* timestamp);
+esp_err_t common_parser_string_to_gnss(const char* sting_in, uint16_t* angle, time_t* timestamp);
 
 esp_err_t common_parser_status_to_string(pivot_config config_in,time_t timestamp, uint16_t angle, char* string_out);
